requirements.txt
atlassian-python-api==1.16.0
PyYAML==5.3.1
